Move serialization structs into own module
We are going to need more code for the transitive orderbook there which
I feel would make the filter module too big.
