+++ price-estimator/src/main.rs
@@ -1,5 +1,6 @@
 mod estimate_buy_amount;
 mod filter;
+mod models;
 mod orderbook;
 
 use core::{
